@@ -1,18 +1,24 @@
 module github.com/duokey/duokey-sdk-go
 
-go 1.16
+go 1.21
 
 require (
-	github.com/coreos/go-oidc/v3 v3.0.0
-	github.com/google/go-cmp v0.5.5 // indirect
-	github.com/google/go-querystring v1.1.0 // indirect
-	github.com/google/uuid v1.2.0
-	github.com/gorilla/schema v1.2.0 // indirect
+	github.com/coreos/go-oidc/v3 v3.9.0
+	github.com/google/go-querystring v1.1.0
+	github.com/google/uuid v1.6.0
 	github.com/pkg/errors v0.9.1
-	github.com/stretchr/testify v1.5.1
-	golang.org/x/net v0.0.0-20201021035429-f5854403a974 // indirect
-	golang.org/x/oauth2 v0.0.0-20210210192628-66670185b0cd
-	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
-	gopkg.in/validator.v2 v2.0.0-20210331031555-b37d688a7fb0
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+	github.com/stretchr/testify v1.8.4
+	golang.org/x/oauth2 v0.17.0
+	gopkg.in/validator.v2 v2.0.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-jose/go-jose/v3 v3.0.2 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/crypto v0.20.0 // indirect
+	google.golang.org/appengine v1.6.8 // indirect
+	google.golang.org/protobuf v1.32.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
